+++ pysat/instruments/pysat_ndtesting.py
@@ -5,6 +5,7 @@ import datetime as dt
 import functools
 import numpy as np
 
+import pandas as pds
 import xarray as xr
 
 import pysat
@@ -33,9 +34,10 @@ concat_data = mm_test.concat_data
 preprocess = mm_test.preprocess
 
 
-def load(fnames, tag='', inst_id='', non_monotonic_index=False,
-         non_unique_index=False, malformed_index=False, start_time=None,
-         num_samples=864, test_load_kwarg=None, max_latitude=90.0,
+def load(fnames, tag='', inst_id='', sim_multi_file_right=False,
+         sim_multi_file_left=False, root_date=None, non_monotonic_index=False,
+         non_unique_index=False, start_time=None, num_samples=864,
+         sample_rate='100S', test_load_kwarg=None, max_latitude=90.0,
          num_extra_time_coords=0):
     """Load the test files.
 
@@ -49,20 +51,27 @@ def load(fnames, tag='', inst_id='', non_monotonic_index=False,
     inst_id : str
         Instrument ID used to identify particular data set to be loaded.
         This input is nominally provided by pysat itself. (default='')
+    sim_multi_file_right : bool
+        Adjusts date range to be 12 hours in the future or twelve hours beyond
+        `root_date`. (default=False)
+    sim_multi_file_left : bool
+        Adjusts date range to be 12 hours in the past or twelve hours before
+        `root_date`. (default=False)
+    root_date : NoneType
+        Optional central date, uses _test_dates if not specified.
+        (default=None)
     non_monotonic_index : bool
         If True, time index will be non-monotonic (default=False)
     non_unique_index : bool
         If True, time index will be non-unique (default=False)
-    malformed_index : bool
-        If True, the time index will be non-unique and non-monotonic. Deprecated
-        and scheduled for removal in pysat 3.2.0.
-        (default=False)
     start_time : dt.timedelta or NoneType
         Offset time of start time since midnight UT. If None, instrument data
         will begin at midnight. (default=None)
     num_samples : int
         Maximum number of times to generate.  Data points will not go beyond the
         current day. (default=864)
+    sample_rate : str
+        Frequency of data points, using pandas conventions. (default='100s')
     test_load_kwarg : any
         Keyword used for pysat unit testing to ensure that functionality for
         custom keywords defined in instrument support functions is working
@@ -90,14 +99,18 @@ def load(fnames, tag='', inst_id='', non_monotonic_index=False,
     drange = mm_test.define_range()
 
     # Using 100s frequency for compatibility with seasonal analysis unit tests
-    uts, index, dates = mm_test.generate_times(fnames, num_samples, freq='100S',
+    uts, index, dates = mm_test.generate_times(fnames, num_samples,
+                                               freq=sample_rate,
                                                start_time=start_time)
-    # TODO(#1094): Remove in pysat 3.2.0
-    if malformed_index:
-        # Warn that kwarg is deprecated and set new kwargs.
-        mm_test._warn_malformed_kwarg()
-        non_monotonic_index = True
-        non_unique_index = True
+
+    # Specify the date tag locally and determine the desired date range
+    pds_offset = dt.timedelta(hours=12)
+    if sim_multi_file_right:
+        root_date = root_date or _test_dates[''][''] + pds_offset
+    elif sim_multi_file_left:
+        root_date = root_date or _test_dates[''][''] - pds_offset
+    else:
+        root_date = root_date or _test_dates['']['']
 
     if non_monotonic_index:
         index = mm_test.non_monotonic_index(index)
@@ -112,7 +125,7 @@ def load(fnames, tag='', inst_id='', non_monotonic_index=False,
     # the root start a measurement is and use that info to create a signal
     # that is continuous from that start. Going to presume there are 5820
     # seconds per orbit (97 minute period).
-    time_delta = dates[0] - dt.datetime(2009, 1, 1)
+    time_delta = dates[0] - root_date
 
     # MLT runs 0-24 each orbit
     mlt = mm_test.generate_fake_data(time_delta.total_seconds(), uts,
@@ -146,6 +159,13 @@ def load(fnames, tag='', inst_id='', non_monotonic_index=False,
     alt0 = 400.0
     altitude = alt0 * np.ones(data['latitude'].shape)
     data['altitude'] = ((epoch_name), altitude)
+
+    # Fake orbit number
+    fake_delta = dates[0] - (_test_dates[''][''] - pds.DateOffset(years=1))
+    data['orbit_num'] = ((epoch_name),
+                         mm_test.generate_fake_data(fake_delta.total_seconds(),
+                                                    uts, period=iperiod['lt'],
+                                                    cyclic=False))
 
     # Create some fake data to support testing of averaging routines
     mlt_int = data['mlt'].astype(int).data