+++ pysat/instruments/pysat_testing.py
@@ -40,8 +40,8 @@ preprocess = mm_test.preprocess
 
 def load(fnames, tag='', inst_id='', sim_multi_file_right=False,
          sim_multi_file_left=False, root_date=None, non_monotonic_index=False,
-         non_unique_index=False, malformed_index=False, start_time=None,
-         num_samples=86400, test_load_kwarg=None, max_latitude=90.):
+         non_unique_index=False, start_time=None, num_samples=86400,
+         test_load_kwarg=None, max_latitude=90.):
     """Load the test files.
 
     Parameters
@@ -67,9 +67,6 @@ def load(fnames, tag='', inst_id='', sim_multi_file_right=False,
         If True, time index will be non-monotonic (default=False)
     non_unique_index : bool
         If True, time index will be non-unique (default=False)
-    malformed_index : bool
-        If True, the time index will be non-unique and non-monotonic. Deprecated
-        and scheduled for removal in pysat 3.2.0. (default=False)
     start_time : dt.timedelta or NoneType
         Offset time of start time since midnight UT. If None, instrument data
         will begin at midnight. (default=None)
@@ -163,13 +160,6 @@ def load(fnames, tag='', inst_id='', sim_multi_file_right=False,
     data['int16_dummy'] = np.ones(len(data), dtype=np.int16)
     data['int32_dummy'] = np.ones(len(data), dtype=np.int32)
     data['int64_dummy'] = np.ones(len(data), dtype=np.int64)
-
-    # TODO(#1094): Remove in pysat 3.2.0
-    if malformed_index:
-        # Warn that kwarg is deprecated and set new kwargs.
-        mm_test._warn_malformed_kwarg()
-        non_monotonic_index = True
-        non_unique_index = True
 
     # Activate if non-monotonic index is needed.
     if np.any([non_monotonic_index, (tag == 'non_strict')]):