+++ pysat/tests/test_methods_general.py
@@ -297,38 +297,3 @@ class TestLoadCSVData(object):
         self.eval_data_cols()
         assert len(self.data.columns) == len(self.data_cols)
         return
-
-
-class TestDeprecation(object):
-    """Unit tests for deprecated methods."""
-
-    def setup_method(self):
-        """Set up the unit test environment for each method."""
-
-        warnings.simplefilter("always", DeprecationWarning)
-        return
-
-    def teardown_method(self):
-        """Clean up the unit test environment after each method."""
-
-        return
-
-    def test_convert_timestamp_to_datetime(self):
-        """Test that convert_timestamp_to_datetime is deprecated."""
-
-        warn_msgs = [" ".join(
-            ["New kwargs added to `pysat.utils.io.load_netCDF4`",
-             "for generalized handling, deprecated",
-             "function will be removed in pysat 3.2.0+"])]
-
-        test = pysat.Instrument('pysat', 'testing', use_header=True)
-        test.load(2009, 1)
-        with warnings.catch_warnings(record=True) as war:
-            gen.convert_timestamp_to_datetime(test, epoch_name='uts')
-
-        # Ensure the minimum number of warnings were raised
-        assert len(war) >= len(warn_msgs)
-
-        # Test the warning messages, ensuring each attribute is present
-        pysat.utils.testing.eval_warnings(war, warn_msgs)
-        return